Frontend. Add EntitySearch component for dynamic autocomplete functionality (#40)
